Rename add-dockings-dialog
